chore: Move the ThemeProvider component
