+++ src/themes/Theme.ts
@@ -1,0 +1,20 @@
+export type Theme = {
+  normal: string;
+  primary: string;
+  secondary: string;
+  error: string;
+  info: string;
+  success: string;
+  warning: string;
+  bg: {
+    primary: string;
+    secondary: string;
+    disabled: string;
+  };
+  text: {
+    primary: string;
+    secondary: string;
+    disabled: string;
+    dividers: string;
+  };
+};
