+++ src/themes/index.ts
@@ -1,1 +1,3 @@
 export * from "./dark";
+export * from "./light";
+export * from "./ThemeProvider";
